src/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-ckeditor==6.1.0
django-embed-video==1.4.0
django-filter==2.4.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
django-social-share==2.2.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.1
urllib3==1.26.5
